% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_account_updates_multi.R
\name{cancel_account_updates_multi}
\alias{cancel_account_updates_multi}
\title{Cancel Account Updates Multi}
\usage{
cancel_account_updates_multi(requests = NULL)
}
\arguments{
\item{requests}{Either a character or a numeric vector containing the names
(\emph{req_name}s) or the IDs (\emph{req_id`s}), respectively, of the
requests that you wish to cancel. Valid \emph{req_name}s and \emph{req_id}s
may be be viewed or retrieved using the \link{subscriptions} environment.}
}
\value{
TRUE upon success (invisible), FALSE or error otherwise.
}
\description{
Stop receiving 3-minute updated Account Update Multi data for an account or
model for which a subscription has been started using
\link{req_account_updates_multi}().
}
\details{
If \emph{requests} is not supplied, then
\code{cancel_account_updates_multi}() will cancel \strong{all} Account
Updates Multi subscriptions.
}
\seealso{
Other treasury: 
\code{\link{cancel_account_summary}()},
\code{\link{req_account_summary}()},
\code{\link{req_account_updates_multi}()},
\code{\link{req_account_updates}()}
}
\concept{treasury}
