% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_mkt_data.R
\name{cancel_mkt_data}
\alias{cancel_mkt_data}
\title{Cancel Market Data}
\usage{
cancel_mkt_data(req_names)
}
\arguments{
\item{req_names}{Character vector containing the names of the subscriptions you wish to
cancel. May contain any or all existing market data subsription names. If
left blank or set to value "**all**", then `cancel_mkt_data`() will cancel
every active market data subscription.}
}
\value{
TRUE upon success (invisible), FALSE or error otherwise.
}
\description{
Cancel an existing market data subscription.
}
\seealso{
Other market data: 
\code{\link{req_market_data_type}()},
\code{\link{req_mkt_data}()}
}
\concept{market data}
