% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_order.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel Specified Orders}
\usage{
cancel_order(orders)
}
\arguments{
\item{orders}{List or char of orders}
}
\value{
* **If SYNC  mode** (`channel` *not specified*)**:**
* **If ASYNC mode** (*specified* `channel`)**:**
}
\description{
Describe
}
\seealso{
Other orders: 
\code{\link{place_order}()},
\code{\link{req_all_open_orders}()},
\code{\link{req_auto_open_orders}()},
\code{\link{req_completed_orders}()},
\code{\link{req_ids}()},
\code{\link{req_open_orders}()}
}
\concept{orders}
