% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancellator.R
\name{cancellator}
\alias{cancellator}
\title{The Cancellator
Handles most `cancel_` API function calls.}
\usage{
cancellator(
  cancel_msg,
  requests = NULL,
  fun_name = rlang::call_name(sys.call(-1)),
  subscription_name = gsub("cancel_", "", fun_name)
)
}
\description{
The Cancellator
Handles most `cancel_` API function calls.
}
\keyword{internal}
