% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_slate.R
\name{clean_slate}
\alias{clean_slate}
\title{Clear out the subscription environments and the sock drawer}
\usage{
clean_slate(what = c("archives", "sock_drawer", "subscriptions", "treasury"))
}
\arguments{
\item{what}{What would you like to clear out?}
}
\description{
Clear out the subscription environments and the sock drawer
}
\seealso{
Other utilities: 
\code{\link{active_connection_parameters}()},
\code{\link{create_new_connections}()},
\code{\link{disconnect}()},
\code{\link{socks}()},
\code{\link{start_api}()},
\code{\link{sync_timeout}()}
}
\concept{utilities}
