% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{client_id_0}
\alias{client_id_0}
\title{Client ID 0: The Trader Workstation Client}
\format{An object of class \code{NULL} of length 0.}
\usage{
client_id_0
}
\description{
A socket that connects to TWS or IBG using Client ID = 0 is considered a *TWS
socket*. TWS sockets have the unique property of being able to view orders
placed within the TWS app itself.
}
\keyword{datasets}
