% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{disconnect}
\alias{disconnect}
\title{Close and Remove an IB Connection}
\usage{
disconnect(
  number_to_disconnect = "all",
  disconnect_master = FALSE,
  disconnect_tws = FALSE
)
}
\arguments{
\item{number_to_disconnect}{Integer. Total number of sockets to disconnect, including Master and TWS
connections if included.}

\item{disconnect_master}{Boolean, default FALSE. If TRUE, then the \link{master_client_id}{Master
Client Socket} will be disconnected if it exists.}

\item{disconnect_tws}{Boolean, default FALSE. If TRUE, then the \link{client_id_0}{TWS Socket}
will be disconnected if it exists.}
}
\description{
The inverse of \link{create_new_connections}(). Will close and remove IB
socket connections as directed.
}
\seealso{
create_new_connections

Other connect and disconnect: 
\code{\link{create_new_connections}()}
}
\concept{connect and disconnect}
