% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{functionary}
\alias{functionary}
\title{The Functionary: A 'dictionary' for functions}
\format{An object of class \code{list} of length 25.}
\usage{
functionary
}
\description{
Coded values needed for communicating with IB's servers over socket.
}
\keyword{internal}
