% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_collate.R
\name{ib_collate}
\alias{ib_collate}
\title{Group multi-response functions into lists.}
\usage{
ib_collate(stamped_responses_list)
}
\description{
Group multi-response functions into lists.
}
\keyword{internal}
