% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_connect.R
\name{ib_connect}
\alias{ib_connect}
\title{Establish a connection with Interactive Brokers}
\usage{
ib_connect(
  master = FALSE,
  tws = FALSE,
  host = active_itr_option(host),
  port = active_itr_option(port)
)
}
\arguments{
\item{host}{(default = \code{localhost}) Where is the IB connection hosted?}

\item{port}{(default = 4002) What port is the IB connection listening on?
Defaults to 4002, which is IB's default port for the paper trading account
port accessible using IB Gateway.}

\item{client_id}{desired client id as an INT or CHAR. Must be a number
between 0 and 32767, inclusive.}

\item{async}{boolean:

"\strong{FALSE}": Creates a socket that will immediately return the server's
  response to an API call, even if it means putting other processes on hold
  while the response is retrieved.

"\strong{TRUE}" (default): The likely case if `ib_connect()` is being called
  by a user instead of within one of the API calls. Will create a persistent
  async socket that, after being writen to, will release controlto other
  processes while the socket retrieves data from IB, which can be read later.}
}
\description{
Creates a socket connection to a running instance of Trader Workstation or
  IB Gateway -- either persistant (async) or one-time (sync).
}
\keyword{internal}
