% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_decode.R
\name{ib_decode}
\alias{ib_decode}
\title{Decode IB Raw Vector}
\usage{
ib_decode(ib_parsed_raw_list)
}
\description{
Accepts a list of raw vector  and returns them as formatted, human
  readable R data objects.
}
\keyword{internal}
