% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_sync_connect.R
\name{ib_sync_connect}
\alias{ib_sync_connect}
\title{IB Sync Connect}
\usage{
ib_sync_connect(
  host = active_itr_option(host),
  port = active_itr_option(port),
  client_id
)
}
\arguments{
\item{host}{Where is the IB connection hosted?}

\item{port}{What port is the IB connection listening on?}
}
\description{
Creates a lightweight socket connection for sync usage.
}
\keyword{internal}
