% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveTradeR-package.R
\docType{package}
\name{InteractiveTradeR-package}
\alias{InteractiveTradeR}
\alias{InteractiveTradeR-package}
\title{InteractiveTradeR}
\description{
API for Interactive Brokers TWS & IBG
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JakeVestal/InteractiveTradeR}
  \item Report bugs at \url{https://github.com/JakeVestal/InteractiveTradeR/issues}
}

}
\author{
\strong{Maintainer}: Jake Vestal \email{jmvestal@gmail.com} (\href{https://orcid.org/0000-0002-2970-4795}{ORCID})

}
\keyword{internal}
