% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveTradeR-package.R
\docType{package}
\name{InteractiveTradeR}
\alias{InteractiveTradeR}
\alias{InteractiveTradeR-package}
\title{InteractiveTradeR}
\description{
Basic info, disclaimer, and overview
}
\section{Disclaimer and Waiver}{

You can use this package for any and all purposes so long as it's connected
to your paper trading account or Interactive Brokers' live demo. You are
violating the license agreement in the instant you connect InteractiveTradeR
to a live Interactive Brokers account without permission. Please contact Jake
Vestal, the package author, for information on licensing.

Not affiliated with Interactive Brokers, LLC in any way.
}

\section{NA}{

You can use this package for any and all purposes so long as it's connected
to your paper trading account or Interactive Brokers' live demo. You are
violating the license agreement in the instant you connect InteractiveTradeR
to a live Interactive Brokers account without permission. Please contact Jake
Vestal, the package author, for information on licensing.

Not affiliated with Interactive Brokers, LLC in any way.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JakeVestal/InteractiveTradeR}
  \item Report bugs at \url{https://github.com/JakeVestal/InteractiveTradeR/issues}
}

}
\author{
\strong{Maintainer}: Jake Vestal \email{jmvestal@gmail.com} (\href{https://orcid.org/0000-0002-2970-4795}{ORCID})

}
