% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_InteractiveTradeR_example_app.R
\name{launch_InteractiveTradeR_example_app}
\alias{launch_InteractiveTradeR_example_app}
\title{Run InteractiveTradeR Example App}
\usage{
launch_InteractiveTradeR_example_app(
  port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  display.mode = c("auto", "normal", "showcase")
)
}
\arguments{
\item{port}{The TCP port that the app should listen on. Defaults to choosing
a random port.}

\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not.}

\item{display.mode}{The mode in which to display the example. Defaults to
\code{showcase}, but may be set to \code{normal} to see the example without
code or commentary.}
}
\description{
Don't forget to open IB Gateway or Trader Workstation first!
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  # List all available examples
  run_example()

  # Run one of the examples
  run_example("01_basics")

  # Print the directory containing the code for all examples
  system.file("examples", package="InteractiveTradeR")
}

}
