% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-orders.R, R/data.R
\name{order_obj}
\alias{order_obj}
\alias{orders}
\alias{order}
\title{Order Objects}
\description{
* **orderComboLegs** <numeric>: A numeric vector whose length equals the
number of rows in the `comboLegs` element in the \link{contract} object.
`orderComboLegs` defines the order prices for each combo leg.

describe
}
\details{
* **smartComboRoutingParams** <named character>: In InteractiveTradeR,
`smartComboRoutingParams` is a named character vector in which a parameter is
identified by a "tag", and its value by "value". The tags and values
available are described in the
\href{https://interactivebrokers.github.io/tws-api/classIBApi_1_1Order.html#a839106698ec5bb1192a22e4e1debdf28}{smartComboRoutingParams}
section of IB's documentation.

For example, to specify a non-gauranteed order, use `smartComboRoutingParams
= c(NonGuaranteed = 1)`.

To instruct the execution system to not start the next leg-in if the previous
leg-in did not finish, you would use `smartComboRoutingParams =
c(DontLeginNext = 1)`

You may specify as many smart combo routing params as you like; simply use
`c`() to combine them into a named character vector; for example:
`smartComboRoutingParams = c(NonGuaranteed = 1, DontLeginNext = 1)`
}
