% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sock_drawer.R
\name{read_sock_drawer}
\alias{read_sock_drawer}
\title{Read Sock Drawer}
\usage{
read_sock_drawer(verbose = FALSE)
}
\arguments{
\item{verbose}{(Boolean, default FALSE): If set to TRUE, `read_sock_drawer`()
will print informative messages to the terminal.}
}
\description{
Read the whole sock drawer.
}
