% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_completed_orders.R
\name{req_completed_orders}
\alias{req_completed_orders}
\title{Request Completed Orders}
\usage{
req_completed_orders(apiOnly = FALSE, channel = NULL)
}
\arguments{
\item{apiOnly}{Logical, length 1, default FALSE. If FALSE, result will include completed
orders submitted via the TWS; if TRUE, then only orders submitted through the
API will be included.}

\item{channel}{One of the following:

\itemize{
  \item \strong{Not Specified} (Default): Opens a new connection to IB, uses
  it to issue the request and retrieve the response, and closes connection
  behind itself upon completion.
  \item \strong{The Name of a Sock}: Character vector, length 1. The name of
  an open, connected socket in the \link{sock_drawer}; e.g., "master", "tws",
  or "sock_123"
  \item \strong{Numeric Client ID}: Numeric, length 1. The client ID for
  which open orders are to be retrieved; e.g., 0, 874, 123. If a client ID is
  passed, and no socket in the \link{sock_drawer} is connected on that ID,
  then a new socket will be opened on that ID, and closed upon function exit.
  \item \strong{A \code{sockconn} Connection}: An open connection object of
  class "sockconn", connected to the IB API; e.g., \code{sock_drawer$tws}
}}
}
\value{
* **If SYNC  mode** (`channel` *not specified*)**:**
* **If ASYNC mode** (*specified* `channel`)**:**
}
\description{
Fetch the orders completed (if any) in the current trading day.
}
\seealso{
Other orders: 
\code{\link{cancel_order}()},
\code{\link{place_order}()},
\code{\link{req_all_open_orders}()},
\code{\link{req_auto_open_orders}()},
\code{\link{req_ids}()},
\code{\link{req_open_orders}()}
}
\concept{orders}
