% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_managed_accts.R
\name{req_managed_accts}
\alias{req_managed_accts}
\title{Request Managed Accounts}
\usage{
req_managed_accts(channel = NULL)
}
\arguments{
\item{channel}{One of the following:

* **Not Specified** (Default): Opens a new connection to IB, uses it to issue
the request and retrieve the response, and closes connection behind itself
upon completion.

* **The Name of a Sock**: Character vector, length 1. The name of an open,
connected socket in the \link{sock_drawer}; e.g., "master", "tws", "sock_123"

* **Numeric Client ID**: Numeric, length 1. The client ID for which open
orders are to be retrieved; e.g., 0, 874, 123. If a client ID is passed, and
no socket in the \link{sock_drawer} is connected on that ID, then a new
socket will be opened using that ID, and closed upon function exit.

* **A `sockconn` Connection**: An open connection object of class "sockconn",
connected to the IB API; e.g., `sock_drawer$tws`}
}
\value{
character vector in which each element is the ID of an account accessible by
the currently logged-in user.
}
\description{
Retrieve the account ID codes for all IB accounts your username can access.
}
\examples{
\dontrun{
  # This example assumes that you're connected to a default PAPER Account
  # having five (5) trading accounts (prefix "DU") and an admin (prefix "DF"),
  # for six (6) accounts total.

  # Fetch managed accounts
  managed_accts <- req_managed_accts()
  # Print managed accounts
  managed_accts
}
}
