% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_market_data_type.R
\name{req_market_data_type}
\alias{req_market_data_type}
\title{Request Market Data Type}
\usage{
req_market_data_type(mktDataType, channel = NULL)
}
\arguments{
\item{mktDataType}{Either:
\enumerate{
  \item String having one of the following values: "\strong{REALTIME}",
  "\strong{FROZEN}", "\strong{DELAYED}", or "\strong{DELAYED_FROZEN}".
  \item Integer from 1 to 4, inclusive, corresponding respectively to each of
  the above choices (e.g., \code{mktDataType = 3} will set the market data
  type to "\strong{DELAYED}").
}}

\item{channel}{One of the following:

\itemize{
  \item \strong{Not Specified} (Default): Opens a new connection to IB, uses
  it to issue the request and retrieve the response, and closes connection
  behind itself upon completion.
  \item \strong{The Name of a Sock}: Character vector, length 1. The name of
  an open, connected socket in the \link{sock_drawer}; e.g., "master", "tws",
  or "sock_123"
  \item \strong{Numeric Client ID}: Numeric, length 1. The client ID for
  which open orders are to be retrieved; e.g., 0, 874, 123. If a client ID is
  passed, and no socket in the \link{sock_drawer} is connected on that ID,
  then a new socket will be opened on that ID, and closed upon function exit.
  \item \strong{A \code{sockconn} Connection}: An open connection object of
  class "sockconn", connected to the IB API; e.g., \code{sock_drawer$tws}
}}
}
\description{
For a specific socket connection, set the market data type to one of
\strong{REALTIME} (initial setting for all new sockets), \strong{FROZEN},
\strong{DELAYED}, or \strong{DELAYED_FROZEN}. In InteractiveTradeR,
\strong{this function should not be called on its own}. Rather, it is used
internally by other \emph{req_} functions, such as \link{req_mkt_data}.
\code{req_market_data_type}() is included in InteractiveTradeR to cover the
hypothetical case in which the user might need it.
}
\section{The Market Data Types}{

The type of market data retrieved on a given sock can be set to one of the
values below. Note that the difference between \strong{REALTIME}, \strong{FROZEN} and
\strong{DELAYED_REALTIME}, \strong{DELAYED_FROZEN} is that the
\strong{DELAYED} data may be obtained for free on an IB paper account without
a data subscription and with no actual funds deposited.
\describe{
  \item{\strong{REALTIME}:}{Streaming market data (live), fed to the socket
  in real time.
  \href{https://www.interactivebrokers.com/en/software/am3/am/settings/marketdatasubscriptions.htm}{Market
  Data Subscriptions} are required to make use of this setting.}
  \item{\strong{FROZEN}:}{Also requires
  \href{https://www.interactivebrokers.com/en/software/am3/am/settings/marketdatasubscriptions.htm}{Market
  Data Subscriptions}. If utilized, this mode will send the last quote
  available for an asset in the event that there is no \strong{REALTIME}
  price available at the time of of a call to \code{req_mkt_data}(); for
  example, if the call is made outside of trading hours.}
  \item{\strong{DELAYED}:}{Does not require any data subscriptions, but will
  return data delayed by 15 to 20 minutes.}
  \item{\strong{DELAYED_FROZEN}:}{Same behavior as \strong{FROZEN}, except
  applies to \strong{DELAYED} data as opposed to \strong{REALTIME} and does
  \strong{not} require any data subscriptions.}
}

Market data types are described in IB's official documentation here:
\href{https://interactivebrokers.github.io/tws-api/market_data_type.html}{Market
Data Types}.

By setting up subscriptions on sockets set to different market types, you may
have \strong{multiple market data subscriptions running simultaneously on
different market types}. This can be handy in use cases for which
\strong{REALTIME} market data is needed for a certain data stream, but
\strong{DELAYED} market data will suffice for a different stream for which a
subscription is not available/desired.
}

\seealso{
Other market data: 
\code{\link{cancel_mkt_data}()},
\code{\link{req_mkt_data}()}
}
\concept{market data}
