% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_matching_symbols.R
\name{req_matching_symbols}
\alias{req_matching_symbols}
\title{Request Matching Symbols}
\usage{
req_matching_symbols(pattern, channel = NULL)
}
\arguments{
\item{pattern}{Character vector of length 1. The search string for which
matching stocks are sought. May be a ticker, e.g. "IBM", or part of a
stock's name, e.g. "International". Not case sensitive.}

\item{channel}{One of the following:

\itemize{
  \item \strong{Not Specified} (Default): Opens a new connection to IB, uses
  it to issue the request and retrieve the response, and closes connection
  behind itself upon completion.
  \item \strong{The Name of a Sock}: Character vector, length 1. The name of
  an open, connected socket in the \link{sock_drawer}; e.g., "master", "tws",
  or "sock_123"
  \item \strong{Numeric Client ID}: Numeric, length 1. The client ID for
  which open orders are to be retrieved; e.g., 0, 874, 123. If a client ID is
  passed, and no socket in the \link{sock_drawer} is connected on that ID,
  then a new socket will be opened on that ID, and closed upon function exit.
  \item \strong{A \code{sockconn} Connection}: An open connection object of
  class "sockconn", connected to the IB API; e.g., \code{sock_drawer$tws}
}}
}
\value{
A \link{tibble} object in which each row corresponds to an asset whose name
or ticker matched the \code{pattern} supplied. The tibble's columns are:
\itemize{
  \item \code{con_id} <chr>: Interactive Brokers' unique numeric ID
  \item \code{symbol} <chr>: Exchange symbol (i.e., ticker symbol)
  \item \code{sec_type} <chr>: Asset class ("STK" for stock)
  \item \code{primary_exchange} <chr>: Financial exchange on which most of
  the asset's trade volume takes place (NYSE, ARCA, etc).
  \item \code{currency} <chr>: Currency in which trades are made.
  \item \code{asset_types} <list>: A list of other financial products
  (options, bonds, warrants, etc) that derive from the asset and are
  available to trade through Interactive Brokers.
}
}
\description{
Search Interactive Brokers' database for stocks (and \code{stocks} only)
 by specifying a search string \code{pattern}. \code{pattern} may be a ticker,
 e.g. "AAPL", or part of a stock's name, e.g. "App".
}
\details{
\strong{Space out calls to \code{req_matching_symbols}() by at least 1
second.} IB will only allow at most 1 call to \code{req_matching_symbols} per
second \emph{per user}, meaning that even if you run seperate searches for
matching symbols on different sockets at the same time, you'll get an error
telling you that IB is busy processing the request for a different client.
}
\examples{
# Fetch matching symbols for pattern "international"
matching_symbols <- req_matching_symbols("international")
# Print matching symbols
matching_symbols

# If that doesn't satisfy, try a few more search strings. Wait 3 seconds after
# each one so as to be kind to IB's servers.
req_matching_symbols("ibm")
Sys.sleep(3)
req_matching_symbols("business")
Sys.sleep(3)
req_matching_symbols("machines")
Sys.sleep(3)

# Say you only wanted to know what asset types -- futures, options, warrants,
# etc -- that were available for the stock traded on the NYSE under the symbol
# "IBM". You could pipe the output of req_matching_symbols() to a series of
# functions with the end result of returning only the info you want:
req_matching_symbols("ibm") \%>\%
  dplyr::filter(symbol == "IBM" & primary_exchange == "NYSE") \%>\%
  dplyr::select(asset_types) \%>\%
  unlist(use.names = FALSE)
}
