% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_open_orders.R
\name{req_open_orders}
\alias{req_open_orders}
\title{Request and Bind Open Orders Placed from Within TWS Itself}
\usage{
req_open_orders()
}
\value{
This function is called for its side effect of updating the
\strong{\code{treasury}}, which takes place every time the function executes.
Additionally, the function's return value depends upon the value passed in as
\emph{return_data} as follows:

\itemize{
  \item \strong{If \code{return_data == FALSE}}: A Boolean success flag,
  returned invisibly, indicating that the function executed correctly and
  updated the \strong{\code{treasury}} with any new data retrieved.
  \item \strong{If \code{return_data == TRUE}}: Any new data retrieved will
  be returned in a \link[tibble]{tibble} in addition to being added to the
  \strong{\code{treasury}}. If no new data is available, returns NULL.
}

\emph{return_data} defaults to TRUE unless \emph{channel} is specified.
}
\description{
This function as implemented in InteractiveTradeR differs slightly from the
default *reqOpenOrders*() function implemented in the standard API in that
`req_open_orders`() takes no arguments as it may only be invoked for
\link[=client_id_0]{Client ID 0}, the TWS GUI client ID.

For requesting open orders on other sockets, use \link{req_all_open_orders}
to fetch all orders, no matter what socket was used to open them, and then
use \href{http://adv-r.had.co.nz/Subsetting.html}{subsetting} to filter the
output \link[tibble]{tibble} by client ID to select the orders you're
interested in.

The only reason to use `req_open_orders`() in InteractiveTradeR is due to an
additional special property of `req_open_orders`(): **when called,
`req_open_orders`() causes any and all open orders that were created within
the TWS itself to become *bound***.

**Definition of "Bound"**:

In IB jargon, an order is "*bound*" if it is assigned an order ID. By
default, orders placed within the TWS do not have order IDs (in other words,
they're *unbound*). Because the API requires order IDs to modify or cancel an
order, a call to `req_open_orders`() is necessary to enable the API to cancel
or modify orders placed within TWS.

Relevant IB documentation may be found on the
\href{https://interactivebrokers.github.io/tws-api/open_orders.html}{Open
Orders} page.
}
\section{Use Case}{

The default behavior of the TWS is to *not* assign order IDs to orders opened
within the app itself. That being the case, consider the situation in which
you or a colleague has opened orders from within the TWI app. Because the
open orders have no IDs, you have no way to identify them in the API, and are
therefore unable to cancel or modify them using the API.

By calling `req_open_orders`(), all open orders that were opened within the
TWS app are returned in a \link[tibble]{tibble} as the function's output.
Additionally, **these orders are given an order ID so that they can now be
modified/canceled within the API**.
}

\section{Use `req_auto_open_orders`() to Assign IDs to Future TWS Orders}{

By calling \link{req_auto_open_orders}, the behavior of the TWS app will be
reconfigured to bind all orders created within the app, but will not affect
open orders that were opened within the app before `req_auto_open_orders`()
was called.
}

\section{Best Practice}{

Make a call to `req_open_orders`() and `req_auto_open_orders`() during the
initialization section of any Shiny app (or script) created with
InteractiveTradeR. This will ensure that:

 1) Open orders that were placed within the TWS app before the
 InteractiveTradeR app was started are assigned an order ID by
 `req_open_orders`() so that they are accessible within the API

 2) Future orders placed within the TWS app while the InteractiveTradeR app
 is running are also accessible to the InteractiveTradeR app because the call
 to `req_auto_open_orders`() reconfigured the TWS to assign them order IDs by
 default.

With this practice, any InteractiveTradeR app only ever needs to call
`req_open_orders`() once during initialization.
}

\section{OPEN_ORDERS and ORDER_STATUSES Treasury Objects}{

This function updates the following \link{treasury} objects:

\strong{OPEN_ORDERS} A \link[tibble]{tibble} in which each row represents an
order that, at the time it was last updated, had not fully filled. The column
names are described in the
\href{https://interactivebrokers.github.io/tws-api/classIBApi_1_1Contract.html}{Contract},
\href{https://interactivebrokers.github.io/tws-api/classIBApi_1_1Order.html}{Order},
and
\href{https://interactivebrokers.github.io/tws-api/classIBApi_1_1OrderState.html}{OrderState}
sections of IB's documentation.

\strong{ORDER_STATUSES} A \link[tibble]{tibble} in which each row represents
an order that, at the time it was last updated, had not fully filled. The
column names are described in the
\href{https://interactivebrokers.github.io/tws-api/interfaceIBApi_1_1EWrapper.html#a17f2a02d6449710b6394d0266a353313}{Order
Status} section of IB's documentation.
}

\seealso{
req_auto_open_orders

Other orders: 
\code{\link{cancel_order}()},
\code{\link{place_order}()},
\code{\link{req_all_open_orders}()},
\code{\link{req_auto_open_orders}()},
\code{\link{req_completed_orders}()},
\code{\link{req_ids}()}
}
\concept{orders}
