% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_global_params.R
\name{sync_timeout}
\alias{sync_timeout}
\title{Set/Get Sync Timeout}
\usage{
sync_timeout(timeout)
}
\arguments{
\item{timeout}{Numeric, length 1, specifying the time in seconds that API
calls made in SYNC mode will wait for a response from IB before giving up.
The default value is 5 seconds.

If not `timeout` not specified, then `sync_timeout`() will fetch the current
timeout setting.}
}
\description{
Set or retrieve the number of seconds that API calls in SYNC mode will wait
for a response from IB before giving up. The default value is 5 seconds.
}
\examples{
# Fetch the current timeout setting for SYNC functions (in seconds):
sync_timeout()

# Set the time out setting for SYNC functions to 10 seconds:
sync_timeout(10)
}
