% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_contracts}
\alias{test_contracts}
\title{Test Contracts}
\format{An object of class \code{list} of length 32.}
\usage{
test_contracts
}
\description{
A list of different types of example contracts.
}
\keyword{datasets}
