% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{`Client ID 0`}
\alias{`Client ID 0`}
\alias{client_id_0}
\title{Client ID 0: The Trader Workstation Client}
\description{
A socket that connects to TWS or IBG using Client ID = 0 is considered a *TWS
socket*. TWS sockets have the unique property of being able to view orders
placed within the TWS app itself.
}
