% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{`Master Client ID`}
\alias{`Master Client ID`}
\title{Master Client ID}
\format{integer between 1 and 65,535, inclusive}
\description{
A particular client ID that the user identifies within Trader Workstation or
IB Gateway. Within \link{InteractiveTradeR}, a socket that connects to TWS or
IBG using the Master Client ID is considered a *master socket*.

The Master Client ID is special becayse any call made \strong{LIST CALLS HERE} , if
made on a master socket, will ---, whereas for a normal one, ---.
}
