% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_connection_parameters.R
\name{active_connection_parameters}
\alias{active_connection_parameters}
\title{Retreive Active Connection Parameters}
\usage{
active_connection_parameters()
}
\description{
Return a \link[tibble]{tibble} that prints nicely and contains the connection
info that InteractiveTradeR is currently using to to interface with the
Interactive Brokers TWS / IBG application.
}
\examples{
# Fetch the current timeout setting for SYNC functions (in seconds):
sync_timeout()

# Set the time out setting for SYNC functions to 10 seconds:
sync_timeout(10)

# Fetch the new setting:
sync_timeout()
}
\seealso{
Other utilities: 
\code{\link{create_new_connections}()},
\code{\link{disconnect}()},
\code{\link{socks}()},
\code{\link{sync_timeout}()}
}
\concept{utilities}
