% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_slate.R
\name{clean_slate}
\alias{clean_slate}
\title{Clear out the subscription environments and the sock drawer}
\usage{
clean_slate(what)
}
\arguments{
\item{what}{What would you like to clear out? Character vector containing any or all of
"archives", "sock_drawer", "subscriptions", "treasury". If \emph{what} is
left blank, the default behavior of \code{clean_slate} is to clear out
each of the listed options.}
}
\description{
This function wipes out data stored in your environments -- use with care!
}
\seealso{
Other utilities: 
\code{\link{active_connection_parameters}()},
\code{\link{create_new_connections}()},
\code{\link{disconnect}()},
\code{\link{read_sock_drawer}()},
\code{\link{socks}()},
\code{\link{start_api}()},
\code{\link{sync_timeout}()}
}
\concept{utilities}
