% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sock_drawer.R
\name{read_sock_drawer}
\alias{read_sock_drawer}
\title{Read Sock Drawer}
\usage{
read_sock_drawer()
}
\description{
Read the every sock that currently exists in the \link{sock_drawer}.
}
\seealso{
Other utilities: 
\code{\link{active_connection_parameters}()},
\code{\link{clean_slate}()},
\code{\link{create_new_connections}()},
\code{\link{disconnect}()},
\code{\link{socks}()},
\code{\link{start_api}()},
\code{\link{sync_timeout}()}
}
\concept{utilities}
