% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socks.R
\name{socks}
\alias{socks}
\title{Socks}
\usage{
socks()
}
\description{
Print a detailed description of all the socks that are currently open.
}
\seealso{
Other utilities: 
\code{\link{create_new_connections}()},
\code{\link{disconnect}()}
}
\concept{utilities}
