% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_api.R
\name{start_api}
\alias{start_api}
\title{Start API}
\usage{
start_api()
}
\description{
In InteractiveTradeR, the API function `start_api`() is implemented
implicitly by other functions such as \link{create_new_connections}();
therefore, there is no need to ever call `start_api`() directly.
}
