% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_timeout.R
\name{sync_timeout}
\alias{sync_timeout}
\title{Set/Get Sync Timeout}
\usage{
sync_timeout(timeout)
}
\arguments{
\item{timeout}{Numeric, length 1, specifying the time in seconds that API calls made in SYNC
mode will wait for a response from IB before giving up. The default value is
5 seconds.

If the argument \emph{timeout} is not specified in the call, then
sync_timeout() will return the current timeout setting.}
}
\description{
Set or retrieve the number of seconds that API calls in \strong{Sync Mode}
will wait for a response from the IB app before giving up. The default value
is set to 5 seconds.
}
\examples{
# Fetch the current timeout setting for SYNC functions (in seconds):
sync_timeout()

# Set the time out setting for SYNC functions to 10 seconds:
sync_timeout(10)

# Fetch the new setting:
sync_timeout()
}
\seealso{
Other utilities: 
\code{\link{active_connection_parameters}()},
\code{\link{create_new_connections}()},
\code{\link{disconnect}()},
\code{\link{socks}()},
\code{\link{start_api}()}
}
\concept{utilities}
